Add create-site-error component to handle errors
While unlikely, the create site endpoint may fail.

In this scenario users are pushed back to the previous onboarding step
without notification of any failure mode ocurring.

This PR adds an interim error step to provide feedback and give users a
way to break out of a potentially broken flow.