G Suite: Fix errors triggered when removing user on Add G Suite page
These errors are displayed when hitting the trash icon when there is only one form to enter a G Suite user. The state of the fields is indeed not reset because the GSuiteNewUser component is never unmounted: since the index used for the list of GSuiteNewUser components is a simple sequence of integers it is not possible to distinguish when the list is reset (the element removed has the same index than the new element). We need to pass a key that uniquely identifies such form in order to avoid that.